// widgets/sellerfinder/src/domain/seller.types.ts
/* -------------------- */
/* Widget Contract      */
/* -------------------- */

export interface Seller {
    readonly id: string;
    readonly name: string;
    readonly category: string;
    readonly lat: number;
    readonly lng: number;
    readonly description: string;
    readonly website?: string;
    readonly rating?: number;
    readonly hours?: string;
}

export type SellerWithDistance = Seller & { distance: number | null };


export interface SellerFinderWidgetConfig {
    readonly data: SellerFinderDataConfig;
    readonly integration?: {
        readonly requires?: readonly GoogleMapsIntegrationName[];
    };
    readonly translations?: TranslationsConfig
}

export interface SellerFinderDataConfig {
    readonly sellers: readonly Seller[];
    readonly defaultCenter: LatLng;
    readonly zoom: number;
    readonly country: string
}

export type LatLng = {
    readonly lat: number;
    readonly lng: number;
};

export type GoogleMapsIntegrationName = 'googleMaps';

/* -------------------- */
/* Runtime              */
/* -------------------- */

export interface ReactEdgeRuntimeConfig {
    readonly integrations: ReactEdgeRuntimeIntegrations;
}

export interface ReactEdgeRuntimeIntegrations {
    readonly googleMaps?: {
        readonly apiKey: string;
    };
}

export type TranslationsConfig = Record<string, string> | undefined;


/* -------------------- */
/* Resolved Config      */
/* -------------------- */

export interface ResolvedSellerFinderConfig {
    readonly data: SellerFinderDataConfig;
    readonly integrations: ReactEdgeRuntimeIntegrations;
    readonly translations: TranslationsConfig
}

export type DistanceUnit = 'km' | 'mi';

export interface DistanceOption {
    label: string;
    value: number;
}

export const BASE_RADIUS_KM = [10, 50, 200, 500];
export const BASE_RADIUS_MI = [10, 30, 100, 300];